feat: onboarding can be now closed
